bugfix 2.version - move useEffect to _app.js
