+++ pages/_app.js
@@ -1,8 +1,10 @@
 import GlobalStyle from "../styles";
 import Head from "next/head";
+import { nanoid } from "nanoid";
 
 import useLocalStorageState from "use-local-storage-state";
 import { vocs, initialSettings } from "../lib/db";
+import { useEffect } from "react";
 
 import Layout from "../components/Layout";
 
@@ -13,6 +15,17 @@ export default function App({ Component, pageProps }) {
   const [settings, setSettings] = useLocalStorageState("voc_settings", {
     defaultValue: initialSettings,
   });
+
+  useEffect(() => {
+    shuffleCards();
+  }, []);
+
+  // Shuffle all cards
+  function shuffleCards() {
+    setCards((prev) => prev.sort(() => 0.5 - Math.random()));
+  }
+
+
   function handleChangeSettings(name, newValue) {
     setSettings((prev) => {
       return prev.map((setting) =>
@@ -20,7 +33,12 @@ export default function App({ Component, pageProps }) {
       );
     });
   }
-
+  function handleAddNewWord(word, translation) {
+    setCards((prev) => [
+      ...prev,
+      { id: nanoid(), word, translation, hits: 0, views: 0 },
+    ]);
+  }
   return (
     <>
       <GlobalStyle />
@@ -32,7 +50,13 @@ export default function App({ Component, pageProps }) {
       <Layout>
         <Component
           {...pageProps}
-          {...{ cards, setCards, settings, handleChangeSettings }}
+          {...{
+            cards,
+            setCards,
+            settings,
+            handleChangeSettings,
+            handleAddNewWord,
+          }}
         />
       </Layout>
     </>
